+++ packages/react-resizable-panels/src/utils/validatePanelConstraints.ts
@@ -19,6 +19,7 @@ export function validatePanelConstraints({
 
     const {
       collapsedSize = 0,
+      collapsible = false,
       defaultSize,
       maxSize = 100,
       minSize = 0,
@@ -33,7 +34,10 @@ export function validatePanelConstraints({
     if (defaultSize != null) {
       if (defaultSize < 0) {
         warnings.push("default size should not be less than 0");
-      } else if (defaultSize < minSize) {
+      } else if (
+        defaultSize < minSize &&
+        (!collapsible || defaultSize !== collapsedSize)
+      ) {
         warnings.push("default size should not be less than min size");
       }